Build CLI item files tool.

+++ utils/download_item_files.py
@@ -18,7 +18,7 @@ def timestamp_to_date_string(timestamp):
     return datetime.fromtimestamp(timestamp / 1000, timezone.utc).strftime("%Y-%m-%dT%H-%M-%S")
 
 
-def main():
+def main(overwrite=False):
     feed_uri = config.FEED_URI  # Destination to put item files to.
     client = ScrapinghubClient(config.SCRAPY_API_KEY)
     project = client.get_project(config.SCRAPY_PROJECT_ID)
@@ -40,17 +40,21 @@ def main():
             # URI to put JSON file to.
             remote_path = urlsplit(feed_uri.replace("%(name)s", spider.name).replace("%(time)s", job_start_time)).path
 
-            # Write items to local file first for better performance.
-            local_path = remote_path.split("/")[-1]
-            with open(local_path, "w", encoding="utf8") as f:
-                json.dump(job.items.list(), f, ensure_ascii=False)
+            file_not_exists = False  # TODO
 
-            # Upload items file to target location.
-            with get_sftp_session() as sftp:
-                sftp.put(localpath=local_path, remotepath=remote_path)
+            if overwrite and file_not_exists:
 
-            # Delete local file.
-            os.remove(local_path)
+                # Write items to local file first for better performance.
+                local_path = remote_path.split("/")[-1]
+                with open(local_path, "w", encoding="utf8") as f:
+                    json.dump(job.items.list(), f, ensure_ascii=False)
+
+                # Upload items file to target location.
+                with get_sftp_session() as sftp:
+                    sftp.put(localpath=local_path, remotepath=remote_path)
+
+                # Delete local file.
+                os.remove(local_path)
 
 
 if __name__ == "__main__":
